pyproject.toml
[tool.poetry]
name = "vector-databases"
version = "0.1.0"
description = ""
authors = ["RutamBhagat <rutambhagat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.1"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
langchain-pinecone = "^0.1.0" # Added langchain-pinecone with compatible Python version
pinecone-client = "^3.2.2"
openai = "^1.23.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
